dip sw for map 90

+++ N8-OS/jumper.h
@@ -9,7 +9,7 @@
 #define	JMP_H
 
 u8 jmpSetup(u8 *path);
-u8 jmpGetSize(u8 map_idx);
+u8 jmpGetSize(u16 map_idx);
 u8 jmpGetVal(u8 *game, u8 map_idx, u8 *val);
 
 #endif	/* JMP_H */
